# libkmod: improve realloc behavior for zstd outbuffer