This more than doubles performance of depmod (and likely modinfo etc.) on my machine.