@lucasdemarchi 2024-07-09
For a modinfo bench I didn't have such a big diff:


```
# Before:
✔ ~/p/kmod/build $ hyperfine -w 1 -r 5 ./bench.sh
Benchmark 1: ./bench.sh
  Time (mean ± σ):     13.838 s ±  2.058 s    [User: 7.485 s, System: 5.642 s]
  Range (min … max):   12.329 s … 17.353 s    5 runs
 
# After:
✔ ~/p/kmod/build $ hyperfine -w 1 -r 5 ./bench.sh
Benchmark 1: ./bench.sh
  Time (mean ± σ):     12.890 s ±  2.374 s    [User: 7.418 s, System: 4.679 s]
  Range (min … max):   11.214 s … 17.065 s    5 runs

✔ ~/p/kmod/build $ cat bench.sh 
#!/bin/bash
set -e

while read f; do ./tools/modinfo $f > /dev/null; done < <(find /lib/modules/$(uname -r) -type f -name *.zst)
```

For depmod I've got 1.031 s ±  0.021 s vs 953.1 ms ±  27.5 ms....

Definitely an improvement, but for me it's more in the ~7% ball park for a real workload.