@q66 2024-07-09
That's due to a specific behavior of the glibc allocator (where directly subsequent reallocs with no allocations inbetween won't always actually allocate, however this comes at a cost elsewhere), but on libcs other than glibc (or with different allocators) it's going to make a much bigger difference