@nekopsykose 2024-07-09
on my machine:  

```
# before:
$ hyperfine -N 'bash bench.sh'
Benchmark 1: bash bench.sh
  Time (mean ± σ):      5.203 s ±  0.022 s    [User: 2.607 s, System: 2.787 s]
  Range (min … max):    5.174 s …  5.244 s    10 runs
$ doas hyperfine -N './tools/depmod 6.9.8-0-generic'
Benchmark 1: ./tools/depmod 6.9.8-0-generic
  Time (mean ± σ):      2.571 s ±  0.013 s    [User: 1.145 s, System: 1.408 s]
  Range (min … max):    2.547 s …  2.591 s    10 runs

# after:
$ hyperfine -N 'bash bench.sh'
Benchmark 1: bash bench.sh
  Time (mean ± σ):      4.103 s ±  0.012 s    [User: 2.364 s, System: 1.934 s]
  Range (min … max):    4.088 s …  4.119 s    10 runs

$ doas hyperfine -N './tools/depmod 6.9.8-0-generic'
Benchmark 1: ./tools/depmod 6.9.8-0-generic
  Time (mean ± σ):      1.481 s ±  0.009 s    [User: 0.887 s, System: 0.576 s]
  Range (min … max):    1.468 s …  1.494 s    10 runs
```

the reason you see much lower values is probably because the allocator in your case is the glibc one, which has a particular behaviour around these reallocs where it gets to skip the whole memcpy in some cases. with other allocators (in our case it's llvm scudo, could also test the stock musl malloc-ng one) this behaviour is not present so you get the following behaviour:

https://share.firefox.dev/4bA5glJ