@q66 2024-07-09
it's not particular to scudo, likely to happen with anything not specifically glibc allocator (having the same quirk as glibc comes with excess memory fragmentation)

but sure, i can add something